KAFKA-6761: Merged code from StreamsTopologyGraph into InternalStreamsBuilder and deleted StreamsTopologyGraph
