+++ streams/src/main/java/org/apache/kafka/streams/kstream/internals/InternalStreamsBuilder.java
@@ -24,24 +24,40 @@ import org.apache.kafka.streams.processor.ProcessorSupplier;
 import org.apache.kafka.streams.processor.internals.InternalTopologyBuilder;
 import org.apache.kafka.streams.state.KeyValueStore;
 import org.apache.kafka.streams.state.StoreBuilder;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.Pattern;
 
 public class InternalStreamsBuilder implements InternalNameProvider {
 
     final InternalTopologyBuilder internalTopologyBuilder;
-    private final StreamsTopologyGraph topologyGraph;
-
-
     private final AtomicInteger index = new AtomicInteger(0);
+
+    private final AtomicInteger nodeIdCounter = new AtomicInteger(0);
+    private final Map<StreamsGraphNode, Set<StreamsGraphNode>> repartitioningNodeToRepartitioned = new HashMap<>();
+    private final Map<StreamsGraphNode, StreamSinkNode> stateStoreNodeToSinkNodes = new HashMap<>();
+
+    private static final String TOPOLOGY_ROOT = "root";
+    private static final Logger LOG = LoggerFactory.getLogger(InternalStreamsBuilder.class);
+
+    protected final StreamsGraphNode root = new StreamsGraphNode(TOPOLOGY_ROOT, false) {
+        @Override
+        void writeToTopology(InternalTopologyBuilder topologyBuilder) {
+            // no-op for root node
+        }
+    };
 
     public InternalStreamsBuilder(final InternalTopologyBuilder internalTopologyBuilder) {
         this.internalTopologyBuilder = internalTopologyBuilder;
-        topologyGraph = new StreamsTopologyGraph();
     }
 
     public <K, V> KStream<K, V> stream(final Collection<String> topics,
@@ -52,21 +68,31 @@ public class InternalStreamsBuilder implements InternalNameProvider {
                                                                          topics,
                                                                          consumed);
 
+        root.addChildNode(streamSourceNode);
+        streamSourceNode.setParentNode(root);
+
         addNode(streamSourceNode);
 
         return new KStreamImpl<>(this, name, Collections.singleton(name), false, streamSourceNode);
     }
 
-    public <K, V> KStream<K, V> stream(final Pattern topicPattern, final ConsumedInternal<K, V> consumed) {
+    public <K, V> KStream<K, V> stream(final Pattern topicPattern,
+                                       final ConsumedInternal<K, V> consumed) {
         final String name = newProcessorName(KStreamImpl.SOURCE_NAME);
 
-        StreamSourceNode<K, V> streamSourceNode = new StreamSourceNode<>(name,
-                                                                         topicPattern,
-                                                                         consumed);
+        StreamSourceNode<K, V> streamPatternSourceNode = new StreamSourceNode<>(name,
+                                                                                topicPattern,
+                                                                                consumed);
+        root.addChildNode(streamPatternSourceNode);
+        streamPatternSourceNode.setParentNode(root);
 
-        addNode(streamSourceNode);
+        addNode(streamPatternSourceNode);
 
-        return new KStreamImpl<>(this, name, Collections.singleton(name), false, streamSourceNode);
+        return new KStreamImpl<>(this,
+                                 name,
+                                 Collections.singleton(name),
+                                 false,
+                                 streamPatternSourceNode);
     }
 
     @SuppressWarnings("unchecked")
@@ -92,6 +118,8 @@ public class InternalStreamsBuilder implements InternalNameProvider {
                                                                                .withTopic(topic)
                                                                                .build();
 
+        root.addChildNode(statefulSourceNode);
+        statefulSourceNode.setParentNode(root);
         addNode(statefulSourceNode);
 
         
@@ -131,6 +159,10 @@ public class InternalStreamsBuilder implements InternalNameProvider {
                                                                                                     .withProcessorParameters(processorParameters)
                                                                                                     .isGlobalKTable(true)
                                                                                                     .build();
+
+        root.addChildNode(statefulSourceNode);
+        statefulSourceNode.setParentNode(root);
+        addNode(statefulSourceNode);
 
         addNode(statefulSourceNode);
 
@@ -185,21 +217,61 @@ public class InternalStreamsBuilder implements InternalNameProvider {
                        stateUpdateSupplier);
     }
 
-    //TODO uncomment and implement when ready
-    public void buildAndOptimize() {
-//        if (topologyNeedsBuilding) {
-//            TopologyOptimizer topologyOptimizer = new TopologyOptimizerImpl();
-//            topologyOptimizer.optimize(topologyGraph, internalTopologyBuilder);
-//            topologyNeedsBuilding = false;
-//        }
+
+    void addNode(final StreamsGraphNode node) {
+        node.setId(nodeIdCounter.getAndIncrement());
+        node.setInternalStreamsBuilder(this);
+
+        LOG.debug("Adding node {}", node);
+
+
+        if (node.parentNode() == null && !node.nodeName().equals(TOPOLOGY_ROOT)) {
+            throw new IllegalStateException(
+                "Nodes should not have a null parent node.  Name: " + node.nodeName() + " Type: "
+                + node.getClass().getSimpleName());
+        }
+
+        if (node.triggersRepartitioning()) {
+            repartitioningNodeToRepartitioned.put(node, new HashSet<StreamsGraphNode>());
+        } else if (node.repartitionRequired()) {
+            StreamsGraphNode currentNode = node;
+            while (currentNode != null) {
+                final StreamsGraphNode parentNode = currentNode.parentNode();
+                if (parentNode.triggersRepartitioning()) {
+                    repartitioningNodeToRepartitioned.get(parentNode).add(node);
+                    break;
+                }
+                currentNode = parentNode.parentNode();
+            }
+        }
     }
 
-    // for testing
-    StreamsTopologyGraph getTopologyGraph() {
-        return topologyGraph;
+    public StreamsGraphNode root() {
+        return root;
     }
 
-    public void addNode(StreamsGraphNode graphNode) {
-        topologyGraph.addNode(graphNode);
+    /**
+     * Used for hints when a node in the topology triggers a repartition and the repartition flag
+     * is propagated down through the descendant nodes of the topology.  This can be used to help make an
+     * optimization where the triggering node does an eager "through" operation and the child nodes can ignore
+     * the need to repartition.
+     *
+     * @return Map&lt;StreamGraphNode, Set&lt;StreamGraphNode&gt;&gt;
+     */
+    public Map<StreamsGraphNode, Set<StreamsGraphNode>> getRepartitioningNodeToRepartitioned() {
+        Map<StreamsGraphNode, Set<StreamsGraphNode>> copy = new HashMap<>(repartitioningNodeToRepartitioned);
+        return Collections.unmodifiableMap(copy);
+    }
+
+    /**
+     * Used for hints when an Aggregation operation is directly output to a Sink topic.
+     * This map can be used to help optimize this case and use the Sink topic as the changelog topic
+     * for the state store of the aggregation.
+     *
+     * @return Map&lt;StreamGraphNode, StreamSinkNode&gt;
+     */
+    public Map<StreamsGraphNode, StreamSinkNode> getStateStoreNodeToSinkNodes() {
+        Map<StreamsGraphNode, StreamSinkNode> copy = new HashMap<>(stateStoreNodeToSinkNodes);
+        return Collections.unmodifiableMap(copy);
     }
 }
